fix: extract pathResolverForMount into platform-specific files for macOS build
pathResolverForMount references overlay.GetOverlayLayers which is
only available on Linux. Move the function to util_mount_linux.go
(with overlay support) and util_mount_unimplemented.go (bind-only).
